:umbrella: Add cypress test for adding an item to a collection
This test is a bit funky. When we add a collector-service in the backend verifying that urls are valid, this won't pass.
We should find a website that generates valid images on demand, or simply a page that generates a page for whichever resource you ask for (though we should rather have an image generator).
However, that will also be a bit funky, as we would depend on that website. What we should do instead is to create a service on the backend to be used for testing, generating different types of users on demand. E.g. a "clean" user, with no items, a normal user, with some items, etc. We could also have an endpoint for generating images on demand. These endpoints could then be called upon from cypress to simulate different situations.