:pencil: Bookmarklet now loads code from our site
To ensure that we could use the bookmarklet during development, a dirty hack is used. A global variable is defined by the bookmarklet, which the code from our servers use to add the item to. This ensures that we can use the bookmarklet during development, by changing the links in the bookmarklet to point towards our beta-site, as that is what we use.